+++ src/builder.rs
@@ -1,6 +1,6 @@
 use crate::{
+    collection::{MenuItemCollection, MenuItems},
     interaction::InteractionController,
-    plumbing::MenuItemCollection,
     selection_indicator::{style::IndicatorStyle, Indicator, SelectionIndicatorController},
     Menu, MenuDisplayMode, MenuItem, MenuStyle, NoItems,
 };
@@ -60,6 +60,20 @@ where
             style: self.style,
         }
     }
+
+    pub fn add_items<I: MenuItem<R, MenuStyle<C, S, IT, P>>>(
+        self,
+        item: &mut [I],
+    ) -> MenuBuilder<IT, Chain<MenuItems<'_, I, R, MenuStyle<C, S, IT, P>>>, R, C, P, S> {
+        item.iter_mut().for_each(|i| i.set_style(&self.style));
+
+        MenuBuilder {
+            _return_type: PhantomData,
+            title: self.title,
+            items: Chain::new(MenuItems::new(item)),
+            style: self.style,
+        }
+    }
 }
 
 impl<IT, CE, R, C, P, S> MenuBuilder<IT, Chain<CE>, R, C, P, S>
@@ -80,6 +94,21 @@ where
             _return_type: PhantomData,
             title: self.title,
             items: self.items.append(item),
+            style: self.style,
+        }
+    }
+
+    pub fn add_items<I: MenuItem<R, MenuStyle<C, S, IT, P>>>(
+        self,
+        item: &mut [I],
+    ) -> MenuBuilder<IT, Link<MenuItems<'_, I, R, MenuStyle<C, S, IT, P>>, Chain<CE>>, R, C, P, S>
+    {
+        item.iter_mut().for_each(|i| i.set_style(&self.style));
+
+        MenuBuilder {
+            _return_type: PhantomData,
+            title: self.title,
+            items: self.items.append(MenuItems::new(item)),
             style: self.style,
         }
     }
@@ -104,6 +133,21 @@ where
             _return_type: PhantomData,
             title: self.title,
             items: self.items.append(item),
+            style: self.style,
+        }
+    }
+
+    pub fn add_items<I2: MenuItem<R, MenuStyle<C, S, IT, P>>>(
+        self,
+        item: &mut [I2],
+    ) -> MenuBuilder<IT, Link<MenuItems<'_, I2, R, MenuStyle<C, S, IT, P>>, Link<I, CE>>, R, C, P, S>
+    {
+        item.iter_mut().for_each(|i| i.set_style(&self.style));
+
+        MenuBuilder {
+            _return_type: PhantomData,
+            title: self.title,
+            items: self.items.append(MenuItems::new(item)),
             style: self.style,
         }
     }